HackerRank Anagrams
Two strings,  and , are called anagrams if they contain all the same characters in the same frequencies. For example, the anagrams of CAT are CAT, ACT, TAC, TCA, ATC, and CTA.

Explanation of the Code:
I found the ASCII number for each char (letter of the String) and added it up into totalA and totalB. Then, I compared totalA and totalB to check if the totals were equal. If true, returns "Anagram". If false, returns "Not Anagrams".